% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashare.R
\name{datashare}
\alias{datashare}
\title{meta multicenter data}
\usage{
datashare(Xa, Ya, M1a, M2a, R1 = 1000, R2 = 1000)
}
\arguments{
\item{Xa}{a dataframe/matrix of multicenter data, including X (concrete / continue)}

\item{Ya}{a dataframe/matrix of multicenter data with the same dim of Xa, including Y (concrete / continue)}

\item{M1a}{a dataframe/matrix of multicenter data with the same dim of Xa, including M1 binary}

\item{M2a}{a dataframe/matrix of multicenter data with the same dim of Xa, including M2 binary}

\item{R1}{bootstrap times}

\item{R2}{permutation times}
}
\value{
vector
}
\description{
meta multicenter data
}
\author{
houlei
}
