% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFTa.R
\name{IFTa}
\alias{IFTa}
\title{the method of instrument confounding variable}
\usage{
IFTa(X, Y, M1, M2, R1 = 1000, R2 = 1000)
}
\arguments{
\item{X}{a vector X (concrete / continue)}

\item{Y}{a vector Y (concrete / continue) with the same length of X}

\item{M1}{a vector M1 (binary) with the same length of X}

\item{M2}{a vector M2 (binary) with the same length of X}

\item{R1}{bootstrap times}

\item{R2}{permutation times}
}
\value{
vector
}
\description{
the method of instrument confounding variable
}
\author{
houlei
}
